requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
openai>=1.0.0
python-multipart>=0.0.5
langchain>=0.1.7
langchain-core>=0.1.7
langchain-community>=0.0.9
urllib3<2.0.0 